+++ src/main/java/mate/academy/dao/BookDao.java
@@ -1,5 +1,6 @@
 package mate.academy.dao;
 
+import mate.academy.lib.Dao;
 import mate.academy.model.Book;
 
 import java.util.List;
