--- src/main/java/mate/academy/dao/BookDaoImpl.java
@@ -1,98 +1,0 @@
-package mate.academy.dao;
-
-import mate.academy.lib.ConnectionUtil;
-import mate.academy.lib.Dao;
-import mate.academy.lib.DataProcessingException;
-import mate.academy.model.Book;
-
-import java.math.BigDecimal;
-import java.sql.*;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-@Dao
-public class BookDaoImpl implements BookDao {
-    @Override
-    public Book create(Book book) {
-        return null;
-    }
-
-    @Override
-    public Optional<Book> findById(Long id) {
-        String query = "SELECT * FROM books WHERE id = ?";
-        try (Connection connection = ConnectionUtil.getConnection();
-             PreparedStatement statement = connection.prepareStatement(query)) {
-            statement.setLong(1, id);
-            ResultSet resultSet = statement.executeQuery();
-            if (resultSet.next()) {
-                String title = resultSet.getString("title");
-                BigDecimal price = resultSet.getBigDecimal("price");
-
-                return Optional.of(new Book(id, title, price));
-            }
-        } catch (SQLException e) {
-            throw new DataProcessingException("Could not establish connection with the DB.", e);
-        }
-        return Optional.empty();
-    }
-
-    @Override
-    public List<Book> findAll() {
-        String query = "SELECT * FROM books";
-        List<Book> books = new ArrayList<>();
-        try (Connection connection = ConnectionUtil.getConnection();
-            Statement statement = connection.createStatement()) {
-            //TODO: better to use regular statement here, need to work it out
-            ResultSet resultSet = statement.executeQuery(query);
-            while (resultSet.next()) {
-                Long id = resultSet.getLong("id");
-                String title = resultSet.getString("title");
-                BigDecimal price = resultSet.getBigDecimal("price");
-
-                books.add(new Book(id, title, price));
-            }
-        } catch (SQLException e) {
-            throw new DataProcessingException("Could not establish connection with the DB.", e);
-        }
-        return books;
-    }
-
-    @Override
-    public Book update(Book book) {
-        String query = "UPDATE books SET title = ?, price = ? WHERE id = ?";
-        try (Connection connection = ConnectionUtil.getConnection();
-             PreparedStatement statement = connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)) {
-            statement.setString(1, book.getTitle());
-            statement.setBigDecimal(2, book.getPrice());
-            statement.setLong(3, book.getId());
-
-            int affectedRows = statement.executeUpdate();
-            if (affectedRows < 1) {
-                throw new DataProcessingException("Nothing was updated ");
-            }
-
-            ResultSet generatedKeys = statement.getGeneratedKeys();
-            if (generatedKeys.next()) {
-                
-            }
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-        return null;
-    }
-
-    @Override
-    public boolean deleteById(Long id) {
-        String query = "DELETE FROM books WHERE id = ?";
-        try (Connection connection = ConnectionUtil.getConnection();
-        PreparedStatement statement = connection.prepareStatement(query)) {
-            statement.setLong(1, id);
-
-            int rowsAffected = statement.executeUpdate();
-            return (rowsAffected > 0);
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
