+++ src/main/java/mate/academy/model/Book.java
@@ -7,6 +7,8 @@ public class Book {
     private String title;
     private BigDecimal price;
 
+    public Book() {};
+
     public Book(Long id, String title, BigDecimal price) {
         this.id = id;
         this.title = title;
